+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Changed
+
+- deletions in upsert
+- format of the Chain Table (block length is now 16)
+
+## [3.1.0] - 2023-03-03
+
+### Added
+
+- add macro to implement Findex traits
+
 ## [3.0.0] - 2023-02-27
 
 ### Refactor